App2/requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.6
google-api-python-client==2.15.0
python-dotenv==0.19.0